// GestionDuStock_UI/GestionDuStockUI/src/app/services/models/sale.ts
/* tslint:disable */
/* eslint-disable */
/* Code generated by ng-openapi-gen DO NOT EDIT. */

import { Customer } from './customer';
import { SaleItem } from './sale-item';
export interface Sale {
  amount?: number;
  customer?: Customer;
  id?: number;
  invoiceNumber?: string;
  saleDate?: string;
  saleItems?: Array<SaleItem>;
}